requirements.txt
Flask
pydub
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
whisper @ git+https://github.com/openai/whisper.git
